INTSTAT-91: Autogenerate arrays with required cmsg-proxy information
This patch adds the initial implementation of the required autogeneration
functionality for the cmsg-proxy. Currently it autogenerates a file for
each .proto file with an array containing cmsg_service_info entries for
each rpc that defines a HttpRule (the code is therefore inherently coupled
with the type definitions in cmsg_proxy.h).

Note that generating a single file with a single array is difficult due to
how the protoc-cmsg compiler is executed. This could be explored further in
the future.

An example of the currently autogenerated file is shown below:

/* Generated by the protocol buffer compiler.  DO NOT EDIT! */

/* Do not generate deprecated warnings for self */

static cmsg_service_info service_info_entries[] = {
  {
    .service_descriptor = &statmond_interface_descriptor,
    .input_msg_descriptor = &cmsg_bool_descriptor,
    .output_msg_descriptor = &return_value_descriptor,
    .api_ptr = &statmond_interface_api_set_enable,
    .url_string = "/v5_4_7/statistics/interfaces/enabled",
    .http_verb = CMSG_HTTP_PUT,
  },
  {
    .service_descriptor = &statmond_interface_descriptor,
    .input_msg_descriptor = &dummy_descriptor,
    .output_msg_descriptor = &cmsg_bool_descriptor,
    .api_ptr = &statmond_interface_api_get_enable,
    .url_string = "/v5_4_7/statistics/interfaces/enabled",
    .http_verb = CMSG_HTTP_GET,
  },
};

Change-Id: I5decd990b57125063762a304ed615234f7eb69bf
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>