INTSTAT-91: Use package name not filename in autogenerated function naming
The filename can have characters that are invalid for C variable naming. Also
the filename can be long and have multiple '/' characters if a long path is
used when passing in the filename. To solve these issues simply use the package
name instead for function naming.

Also use MakeHeaderDefineFromFilename to generate the #ifdef directive.

Change-Id: I877e2ecc3fc84c3abd6706df953d4e7b73bf7bd7
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>