INTSTAT-84: Add cmsg_proxy_init function
This function reads in all of the autogenerated proxy definition
libraries (that it is told about) and puts the information into a
GList so that it can be easily accessed later.

In the future the GList should be replaced with a hash table for each
HTTP verb. For now while we are just trying to get something working and
the number of APIs is small a GList is OK to use.

Change-Id: I613b922a69871f51fbaec35e9d491506453b74bd
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Justin Jose <justin.jose@alliedtelesis.co.nz>
Reviewed-by: Pieter Britz <pieter.britz@alliedtelesis.co.nz>