INTSTAT-87: Add _cmsg_proxy_create_client() function
Add a function to create a CMSG client using a CMSG service descriptor
and add this client to a new list which holds all of the required
CMSG clients for the proxy.

Note that in the future this code will be updated to intentionally
only support CMSG servers that use unix sockets. For now use TCP
as statmond uses that for now.

Change-Id: I2d066ef2c53ca2312439652e7c1f15b38c3d3209
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Justin Jose <justin.jose@alliedtelesis.co.nz>