INTSTAT-87: Fix memory leak in _cmsg_proxy_create_client()
If cmsg_client_new() fails and returns NULL then currently the memory
allocated for the transport is leaked. A unit test was written to
expose this leak in a test and then the fix was tested to make sure
the test no longer showed a leak.

For some reason valgrind does not pick up the leak if the function
is only run once in the test...

Change-Id: I506087b39b4450071aa3838da1bd9295188050bf
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Pieter Britz <pieter.britz@alliedtelesis.co.nz>