INTSTAT-122: Manually merge protobuf-c source/header files against v1.0.0 upstream
Manually update protobuf-c.c and protobuf.c-h to the v1.0.0 upstream code.

Note that a pre-existing bug in an ATL_CHANGE in repeated_field_pack() is fixed
in this patch. Specifically a break at the end of the case statement for
PROTOBUF_C_TYPE_UINT8 appears to be missing.

Also note that the default memory allocator has been removed in the upstream code.
Subsequently NULL is now passed as the allocator pointer to ensure the default
system functions (malloc/free - which we manually override with our own memory
tracking functions) are used.

Change-Id: Ie7e3fd7a8fffafcdc5e6ffe5092286b7ccf69961
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>