INTSTAT-124: Create GNode N-ary tree to store URL string
CMSG proxy needs to support URLs that have parameters in the path.
For this we need to have a GNode tree that has all the URLs added
to it. URLs are taken from 'cmsg_service_info', tokenized and used
to create the tree.

A new function is introduced to tokenize the URLs and populate the
tree. The existing 'proxy_entries_list' will be cleaned up later once
the tree implementation is completed.

Test cases are added to test the new function. Added Novaprova setup function.

Change-Id: I72a0d43d79a4b83e2149dd80b966defece07e55b
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Justin Jose <justin.jose@alliedtelesis.co.nz>