INTSTAT-153: Update cmsg-proxy to use UNIX socket transports
It will be a (documented) requirement that any API that wants to be
exposed to the cmsg-proxy must use a UNIX socket transport. This is
because UNIX sockets are more efficient and secure than TIPC/TCP
sockets. Simply update the cmsg-proxy to use UNIX socket transports.

Also update the unit tests to account for this change.

Change-Id: I177993c4c099b517c866e5d4c766eb57346c5cca
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>