INTSTAT-124: Add cmsg_service_info ptr to the leaf nodes
Once the URL string is added to the tree, cmsg_service_info
need to be added to the last nodes. A new structure is defined
to store all the 5 CMSG HTTP actions. If the CMSG API info
node doesn't exist, it is created at the first time.
This service info will be used later to serve user requests
The URLs in the user request is parsed and traversed through
the tree. When there is a match found, cmsg_service_info
in the leaf node used to perform the corresponding action.
'cmsg_proxy_api_info' node is inserted as the first child.

_cmsg_proxy_add_service_info() is changed to take new parameter
cmsg_service_info. 'url' is removed from the argument list because
it can be taken from the new parameter.
Also unit test is changed to pass the new argument. New RPC with
a different URL is added into the test proto file.

Change-Id: I92519a5989d9dcda4951c0ca2f2b97047aaeb231
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Justin Jose <justin.jose@alliedtelesis.co.nz>