VLANAPI-31: Ensure loopback clients are at the end of composite child lists
The whole point of using a composite client is to gain as much efficiency
as possible when executing cmsg APIs across a stack (i.e. by executing in
parallel rather than serially). If the loopback client is at the start of
the child list then we must wait for the invoke_send() of the loopback
client (which actually executes the impl function) before simply sending
the messages across the sockets for each non-loopback client. This
severely reduces the performance of the composite client.

Therefore, always execute the invoke_send() of loopback clients last
to ensure that all other stack members are executing the impl while
the local node is executing the impl via the loopback.

Change-Id: Icec910b7c8090c64fe1407653d1101275176ddaa
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>