INTSTAT-173: Use dlopen to dynamically load proxy definition libraries at startup
To remove the dependency on the 'ipc' package and the need for HAVE_STATMOND
we can simply load the associated libraries at runtime if they exist. This
patch simply replaces the current functionality by dynamically loading the
libstatmond_proto_proxy_def.so library. A future patch will make this generic
to load any *_proto_proxy_def.so library that exists.

Tests:
 - The statmond API still works on arc targets
 - The unit tests all still pass
 - The x510 (which does not have this library) functions as expected (does not crash).

Change-Id: I7a3b3a403512d2321412158a8a06fbdbf35fcaec
Reviewed-by: Justin Jose <justin.jose@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>