INTSTAT-173: Don't leak dynamically loaded library handles
We can't close the handles while the cmsg-proxy library is being used
(as the memory locations from the loaded libraries need to be accessed
for the library to work), however we should not leak the handles.

Simply store the handles in a list and call dlclose on the handles when
the deinit function is called.

Change-Id: I83347a3b0c43287b2997d4309879fced4b3596b8
Reviewed-by: Justin Jose <justin.jose@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>