VLANAPI-98: Present cmsg_proxy errors in the ant api result format
When an error is handled within the CMSG proxy, such as invalid argument
parsing, format the result presented to the API as an ant api result message,
consistent with the rest of the API.

Remove null input check from _cmsg_proxy_call_cmsg_api, as for a message
with all optional fields null input is valid. The protobuf2json library
handles the case where required fields are missing.

Change-Id: Id6862aeadbc9c07c7e3644c91c9b45478ce01bf7
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>