VLANAPI-149: Don't remove ant code from returned messages
Now that the code string has been removed from the returned messages, the
code value needs to be left.  As it is now defined as an enum, the
protobuf2json code will convert it to the equivalent string.

As the field is now named code, the resulting output will be identical to
before.

Change-Id: I812789c49bd55dd949041ed73dd98a1486c92a7c
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>