INTSTAT-210: Add developer-debug code to see mem info in cmsg proxy
This patch adds a debug routine to check memory allocation info in cmsg proxy.
Also to be useful dump the info to a file upon receiving SIGUSR2.

CMSG memory tracing is also turned on as it's also useful.

Note that this code is compiled only if CMSG_PROXY_MEM_DEBUG is defined
during compile time.

Change-Id: If194ec8c977b67de40f057b87af9d947902fa059
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>