VLANAPI-184: Handle the case that json_loads() fails
In function _cmsg_proxy_json_object_create(), the case is not handled if an error
occurred in the calling of json_loads(). This may cause unexpected results from
the subsequent cmsg proxy operations.

This patch updates _cmsg_proxy_json_object_create() so it will pass back error
occured in the calling of json_loads(). And in cmsg_proxy(), it handles the case
if json object creation fails.

Change-Id: I035713936ad5bda22f888e074c820f9d6ae0dcd6
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Stan Xiang <stan.xiang@alliedtelesis.co.nz>