CR-56557: Refactor CMSG packet creation out of cmsg_client_invoke_send()
To allow the queueing to be split out from cmsg_client_invoke_send() we
must first refactor the code that actually builds the CMSG packet. Note
that this new function could also be used in other parts of the CMSG code
base (as it is duplicated in many places).

No functional changes are intended.

Change-Id: If2bd4241fdd94f0ebe6fc27801a6ceeeb7a8ab20
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>