CR-56557: Change publisher to simply invoke the client
With the knowledge that the clients a publisher uses are one-way, and
that the invoke of the child handles queueing the messages if required,
simply change the publisher to invoke each child. Using the function
pointer, rather than the function itself, also makes it clearer that
the publisher is built on top of the cmsg client.

Change-Id: I3da750ba68bf39670a2e2efa1d3de89a0be2b7d1
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>