ENGBL-736: Move connection from client/server onto transport
The idea of a 'connection' is based around sockets that can be connected
to one another. Considering these sockets are at the transport layer and
that the bulk of the code that uses the connection information is in the
transport layer it makes sense to keep this information with the transport.

Change-Id: I28f8a6f05e1777025fd476311ee4f855efa97359
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>