ENGBL-736: Add CMSG_LOG_TRANSPORT_ERROR macro
Add a new macro to specifically log errors that occur in the transport
layer. Update any code in the transport layer that currently uses the
macros CMSG_LOG_SERVER_ERROR/CMSG_LOG_CLIENT_ERROR to instead use
the new CMSG_LOG_TRANSPORT_ERROR macro.

To ensure that valuable debug information is not lost, ensure we store
the object id of the parent (client/server) using the transport so that
it is printed as part of the transport message. Note that a publisher is
currently never passed into the transport layer, it is always the server
or client of the publisher. Therefore, cmsg_transport_write_id() is a
good place to implement this functionality.

Change-Id: If40dc47672cb9b9c5b59c4605646c1721af09162
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>