CR-56676: Add macro for enum check
For enums, checking only the has field is not enough as the
value could be zero which indicates not set. For enums, we
only need to check the value is not set to 'not set'. i.e. not
zero. This is because if the field is not present it gets set
to zero anyway. Not using the 'has' field also gets us ready for
proto3 as that doesn't have 'has' fields.

Change-Id: I4ba9a786273222637703c44822ec5894ad11109f
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Samuel Varley <samuel.varley@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>