CR-56720: Remove outer key of returned JSON under specific circumstances
Under certain circumstances it is prefered if the returned JSON to the
web API caller has the key removed.

One example is when the key is the last part of the template URL, i.e.

GET /v0.1/ppp/interfaces/{if_name}/username

instead of returning

{
   "username": "manager"
}

it makes more sense to simply return

"manager"

This patch also adds similar logic to check for keys named "_data"
and "_error_info" (these words will now be reserved keywords in terms
of cmsg/cmsg-proxy).

Change-Id: I1e68f554123ee438510adbeeb8fe6dfeaf85a4f1
Reviewed-by: Gregory Hoggarth <gregory.hoggarth@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>