CR-56720: Support JSON values with PUT/POST API requests
In some cases we should allow just a JSON value to be input with PUT/POST
API requests. This is when the API is only expecting one JSON value to be
input. To work this out in the code, simply compare the number of fields
parsed from the URL and the number of fields in the protobuf-c message
descriptor. Even though the field will not appear in the documentation,
care should be taken around the '_error_info' field that may potentially
be in the message descriptor.

Change-Id: I6ec8d54cea6133e1a9904620181bb443df2f343f
Reviewed-by: Sam Stephenson <sam.stephenson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>