DNSAPI-69: Add support for parsing query-strings
This patch adds support for parsing query-strings in requests to the
Web API. The query strings are added to the same list of key-value
pairs that the URL parameters are added to, and then validated
against the input message type. Support has been added for validating
the protobuf 'bool' type.

The CMSG API always expects a non-null input (otherwise it asserts).
If a Web API expects no input JSON and no URL parameters, but can
take optional query parameters, then if the user supplies no query
parameters (which is valid) then currently no protobuf message will
be created.

To solve this, if no input JSON is provided, create an empty JSON
object. This is so that an empty protobuf message will be generated to
be passed to the API. The impl function can work out that no fields
were actually set in the input message.

Change-Id: Ic9e68cb1a049bd02a5db086c5e84f11fd72cf952
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>