CMSGPROX-85: Only strip error_info for GET API calls
Currently if the ant code in a result is ANT_CODE_OK, cmsg-proxy strips the
ant_result structure out of the returned message. This is to remove it's
unnecessary presence from gets. However, when modifying state, it can
potentially still include useful information (like CLI_SUCCESS_PLUS_INFO).
This is also required to support bulk operations where the operation is
partially successful and we are returning HTTP code OK, but the response
details still need to be examined.

Change-Id: Ica62b534c6eb196483c24bf08dd231e8743e84cc
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>