ENGBL-736: Loopback client can not use same transport for internal server
Internally the loopback client has a cmsg server which it uses to invoke
the impl function directly in the process space. Currently it shares the
transport between the client and the internal server, however after the
patch:

commit db78965 ("ENGBL-736: Move connection from client/server onto transport")

It is no longer possible for the loopback to use the same transport for
both the client and server as a distinct connection is required for each client
and server.

i.e. the following lines of code do not work when the transport is shared:

/* client uses the pipe's read socket, server uses the write socket */
client->_transport->connection.sockets.client_socket = pipe_fds[0];
server->_transport->connection.sockets.client_socket = pipe_fds[1];

Change-Id: Ida0d2259026b481f1463e5660b5debd419cd2664
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>