CMSGPROX-95: Write unit test for query strings
This patch adds a unit test for the query strings functionality in
cmsg-proxy. The order in which the url parameters and the query
parameters are parsed is important, so the relevant functions calls
have been moved into a new function.

The test tests whether the query string is correctly parsed and the
parsed parameters do not overwrite parameters extracted from the URL
path. Since a new RPC has been added to cmsg_proxy_unit_tests.proto,
the test_cmsg_proxy_service_info_init__list_length test has been
updated to check for the correct number of nodes.

Change-Id: Ia0ca368b54b0a1ed9c849ff916a34e3d09264702
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>