MAINT: Leaf nodes in the cmsg-proxy tree always have data
The leaf nodes in a correctly constructed cmsg-proxy tree will always
have non-NULL data on the leaf nodes (i.e. the data required to map
from the URL to the CMSG service). There is therefore no need to check
for NULL, and the existing comment is wrong, we would never want to end
the tree traversal early as this would cause a memory leak.

The unit test coverage confirms this code is never hit currently.

Change-Id: I67d4f2241d500fc2408c7b854eb07d8cd76611de
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>