MAINT: Non-leaf nodes in the cmsg-proxy tree always have data
The non-leaf nodes in a correctly constructed cmsg-proxy tree will always
have non-NULL data on the non-leaf nodes (i.e. the tokenised string for each
part of the URL). There is therefore no need to check for NULL, and the
existing comment is wrong, we would never want to end the tree traversal
early as this would cause a memory leak.

The unit test coverage confirms this code is never hit currently.

Change-Id: I3b73ecd9b61efe0db9958eea1f1b5b502cd50683
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>