MAINT: Add functional tests for CMSG proxy
Add the ability to implement functional (end to end) tests of the
CMSG proxy functionality. For now simply put them in the unit tests
executable however in the future it may make sense to create a separate
executable for these functional tests.

Note that the 'test_get' RPC is updated to return 'ant_result_plus_bool'
because all APIs that go through the proxy must include the 'ant_result'
message.

Change-Id: Id8e5cf50df35afe14f10154f7e1792a377b7854a
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>