MAINT: Remove 'cmsg_proxy' package from functional tests .proto file
Since these are only used for the functional tests there is no chance of them
being used elsewhere (and needing to use the package name as a 'namespace' for
the RPC/message names). Currently the lines in the tests are getting very long,
we can reduce these by not using a package name for the autogenerated code.

Change-Id: I44d723761f668324c08ee0a30bd8e924264a3611
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>