MAINT: Add functional tests for valid JSON value/array to JSON object conversion
The CMSG proxy allows messages that only contain a single key-value pair to simply
be input as a JSON value/array (i.e. not an object with the key included). Internally
this is then converted to the required JSON object before being converted to protobuf.

Add functional tests to test this functionality for valid requests.

Change-Id: Ic67ea42cef3dbd5da77b01ee9253c45f05f0c609
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>