DNSAPI-159: Implement RPC body mapping functionality
This patch implements the functionality for specifying the RPC input
body mapping. The implementation makes use of the HttpRule body field,
and matches how it is used by GRPC.

Support for mapping the body (or URL parameters) to nested fields has
not yet been added. Support may be added in a future patch.

The OpenAPI documentation already interprets the body mappings
correctly.

Functional tests have been added for the following cases:
- mapping the body to a sub-message field
- mapping the body to a primitive field
- mapping the body to multiple remaining fields that are not URL
  parameters, using "*"
- mapping the body to a single remaining field that is not a URL
  parameter, using "*"
- no body mapping

These tests call the respective CMSG API and in the IMPL function
check that the fields in the input message have been set correctly.

The following functional tests have been modified to specify a body
mapping:
- test_single_bool_put
- test_single_string_put
- test_single_uint32_put
- test_single_repeated_uint32_put

Change-Id: I726b09b13bade1bea899c840f3ddd2e7abfaa3ea
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>