MAINT: Split service tree functionality into a separate file
Split this code out into a new file to improve modularity and maintainability.
For now simply extern a couple of variable and remove static from all the
required functions and put them in the header file.

A future patch will remove the need for the extern variables and reduce the number
of functions declared in the header file.

Change-Id: Ie6473bce94ab7cfa13d05331e824ef681af4c2f0
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>