CMSGPROX-118: Ensure only '_error_info' field is returned on error
Currently if there is data set in a message that is returning something other
than ANT_CODE_OK this data will be returned in the JSON dependent on whether
there is more than 2 fields in the message (i.e. it behaves as we wish when
there are two fields, however anymore then the data is returned).

Simply update the logic in _cmsg_proxy_generate_json_return() to properly implement
the logic we desire.

Functional tests that were previously commented out are now enabled and pass. All
existing functional and unit tests also still pass.

Change-Id: I2ace7797b49a18d9db8349ecda61aacd9faea006
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>