CR-57427: Strip details array from ant_result JSON if it is empty
With the change to the protobuf2json-c library NULL repeated fields
are now correctly converted to empty JSON arrays as per the specifications
outlined for proto3. This however makes the ant_result message returned
from most APIs rather ugly as they now all include an empty "details"
array. To solve this specifically just for the ant_result message, simply
delete this field if it is an empty array.

Note that _cmsg_proxy_protobuf2json_string() is now unused and can be deleted.

Change-Id: I2e3ae73d5eba067a210cb64d3e5015228ffe5de9
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>