CR-56223: Don't shutdown/close client socket on server transport
The socket stored on the client socket of the server transport connection
is transient and is simply set as API requests are received on the server.
This socket is used for sending the response back to the API caller. As such
the code calling the cmsg_server_receive() function is in charge of closing
this socket. Avoid a potential double close of the socket descriptor by removing
the existing code in the transport server close function.

Change-Id: I14f70dbf73c7b39af3bbfdd71cacc3f9d323e1a5
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>