ENGBL-832: Delete cmsg-test application code
A Novaprova test suite for CMSG functionality has been implemented
in the directory /cmsg/test. Since this test suite provides much
better coverage (with which we can see using LCOV/GCOV) and is easier
to maintain we no longer require the cmsg-test application.

Simply delete this code to reduce the amount of code to maintain in
the cmsg repo. It is always in the git history in the unlikely case
we ever wish to recover the code.

Change-Id: I3751e5554bc8d7e71a98ef1eda0de1544c1d0d4f
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>