MAINT: Remove profiling code
The profiling code has not been run for years and likely does not
even compile. The code simply adds clutter (the macros are throughout
the core code) and make the code even harder to read.

Simply delete this code to reduce the amount of code to maintain in
the cmsg repo. It is always in the git history in the unlikely case
we ever wish to recover the code.

Change-Id: Ib6ffc1c95dacd5decddebc81549e8d9205b16012
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>