MAINT: Remove local-install script
The CMSG library has undergone significant refactor in the previous
year. One such refactor is that we now build the protobuf-c library
separate to CMSG. To build CMSG locally you had to tweak the source
code just to get it to build. Now that the protobuf-c library is built
separate (using the upstream code) building the library locally cannot
be done using this script.

Simply delete this code to reduce the amount of code to maintain in
the cmsg repo. It is always in the git history in the unlikely case
we ever wish to recover the code.

Change-Id: I7e99917bf3de8272d1221a3e4812359002fd9288
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>