MAINT: Update impl stub generator to include send_msg init and send
Also remove the comments stating that the code is temporary and rename
the file.

This will simplify creating new CMSG APIs by providing the initialization
and send calls so that they don't have to be manually typed or copied
and edited from another IMPL function with the possibility for errors.

This is especially useful for generating web APIs as a lot of APIs are
being generated with different send/recv messages.

Change-Id: I050a6196188c012cf3fb2322e56388048d23b144
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>