SDWAN-201: Fix memory leak in loopback_client_recv
When a large message was being received, a new
buffer was allocated but this was not being freed
in the successful case.

This was found in a project stream and is being
applied to main before delivery.

New tests have been added to test this for all
transports.

Removed the if checks for recv_buf as freeing
a NULL pointer is safe.

Change-Id: Ied97cb8657a4113f8e14333a497002b757549655
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>