MAINT: Infer publisher queuing enabled from queue length
Currently the boolean containing whether queuing is enabled on a publisher
or not is not set in all possible paths. This means that potentially queued
entries are not free'd on unsubscription and previously free'd memory is used
when processing the queue later on.

Rather than updating each path, simply infer whether queuing is enabled and we
need to free potentially matching entries based on if there is anything in the list.

Change-Id: Ib99fcac0eab0e4d83a2ec1e4b6a5872f4a38f495
Reviewed-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>