ENGBL-736: Check for calloc() failure in _cmsg_transport_server_recv function
This really should already being checked anyway, but update the server_recv
to be the same as the client_recv in that it checks if the calloc() worked
or not. Also modify the client_recv so that it no longer closes the transport.
This is being done in the client code anyway on failure and the transport layer
should not be making the decision to close itself anyway (leave this to the
parent client/server).

Change-Id: Ica4e6860e98d984e5c77c8875d62f136f80a90d9
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>