CR-58077: Ensure 'cmsg_pub_subscriber_remove_all' decrements subscriber count
Currently the subscriber count is not decremented when removing each entry
from the list. This has potential to cause the count to be incorrect if the
publisher is then re-used after the function is called. Update the function
to use '_cmsg_pub_subscriber_delete_link' which will ensure the count is
decremented correctly.

Change-Id: Iedf74a3aa1a5859239e7e0466b6e3fa3a266c0b9
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>