CR-58077: Don't return early if subscriber_count is zero during retry period
The only way subscribers can unsubscribe during the retry period is if the
sending of the publisher is done in a separate thread to the thread that
handles subscribe/unsubscribe requests. In this case unsubscribe will handle
freeing any required entries off the queue so this code here is redundant.

Change-Id: Ie04b4e12c3c34b693161e34a99446b0a3509cc6f
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>