INTSTAT2-131: Fix threading issues of composite client
There are two issues with the composite client that mean that it is not thread safe.

The first issue is that the lock on the composite client itself is unlocked once all
of the sending is done on the child clients. The unlock should instead be once all
the receiving on the child clients is completed. Otherwise the child clients may have
more data on their sockets from a composite client invoke in another thread before
the current thread reads them.

The second issue is that the child clients themselves are not locked while the composite
client is invoked. This means that the child client can be invoked from two threads
concurrently. The fix for this is to lock the child clients similarly to how they are
locked when invoked standalone via cmsg_client_invoke.

Change-Id: I2064bbdf5d15e135bfcb0ed71d0bc6d7441307b0
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Justin Jose <justin.jose@alliedtelesis.co.nz>