INTSTAT2-137: CMSG client with loopback causes problem with empty message
When RPC with loopback transport receives an empty message, it leaves some
data in the socket and returns early. This causes problem for the next message.

This problem exists only in loopback transport. In other transports, the full
header is cleaned up properly.

Also added unit tests to cover this issue which shows the problem is detected
in the original code.

Change-Id: I81f908e330d02c622e4302e0d875df4dfbc79034
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>