CR-57878: Don't use transport id in CMSG counters app name for UNIX transport
Currently we create counters everytime a client/server/publisher/subscriber
is created. Usually these just re-use existing (persistent) counters as the
name which includes the transport id matches (i.e. TIPC and TCP always use
the same port numbers). This falls over however when UNIX based subscribers
continuously come and go which is the case for IMISH sessions that subscribe
for ATMF notifications. Since the transport ID for these subscribers uses the
process id (PID) eventually we run out of counterd applications to use and
the error logs for this CR are seen.

The fix is to remove the transport ID from the counterd application name when
the transport is of UNIX type. That way all of the UNIX based subscribers
used by IMISH sessions will all use the same application.

Change-Id: Icdb5fc0b9e5742a22066d0835758de51adbe17c5
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>