ENGBL-835: Crypto functionality now implemented in ATMF as UDT
This functionality is now implemented in ATMF as part of a UDT. As such
this code can now be deleted from CMSG as it is implemented there. There
will be another patch in the future to delete the rest of the crypto
structures from CMSG.

Note: The addition of cmsg_transport_server_recv_process_DEPRECATED is
required to simply move the existing crypto functionality out into ATMF
as a UDT. Mark it as deprecated to ensure no one tries to use this in
any new code.

Change-Id: I178597ca685194f17faf0b4bcf34f9c19c3110b6
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>