CR-58429: Fix error logs when stack separates
The HSL_MUX_OVERALL_RETVAL_RESULT is designed to silently ignore
CMSG_RET_CLOSED errors, but the cmsg code for the composite client was
actually overwriting any return code with CMSG_RET_ERROR if it failed
during the send. This caused errors to be printed when a stack was
separated and the timing happened to be such that the MUX call was done
before the child was removed from the composite client.

The fix is to use the actual return code and return CMSG_RET_CLOSED
when the connect fails.

Change-Id: I253ee13f1e30e0c92bde9f83608def7b4fd8d46e
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>