ENGBL-736: Remove (void *) casting from recv_wrapper functions
Now that we are passing in the transport to the recv_wrapper we can simplify
this code by passing the socket directly without casting it.

Note that for the sake of reviewing this patch, the 'socket_dup' field is
removed in the next patch (it duplicates an already passed in parameter).

Change-Id: Ic578b46547e9c5ba21d4678a9afd828f2a550012
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>