ENGBL-736: cmsg_client_invoke_recv_direct is redundant
We can simply call 'cmsg_client_invoke_recv' like all the other transports
that receive a reply. 'cmsg_client_invoke_recv_direct' is simply a cut down
version of 'cmsg_client_invoke_recv' and makes a distinction between receiving
on a loopback transport and all others, when in fact the receiving is exactly
the same. The fact we are receiving off a pipe (instead of a socket) is abstracted
by the transport layer.

Change-Id: I6ccb6bc318b890ada52ac77693d24d1d4cef1427
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>